services/image-detector/requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python-headless
Pillow
requests